+++ src/main/java/org/truffleruby/core/proc/ProcNodes.java
@@ -27,6 +27,7 @@ import org.truffleruby.core.string.StringNodes;
 import org.truffleruby.core.symbol.SymbolNodes;
 import org.truffleruby.language.NotProvided;
 import org.truffleruby.language.Visibility;
+import org.truffleruby.language.WarnNode;
 import org.truffleruby.language.arguments.ArgumentDescriptorUtils;
 import org.truffleruby.language.arguments.ReadCallerFrameNode;
 import org.truffleruby.language.arguments.RubyArguments;
@@ -62,7 +63,6 @@ public abstract class ProcNodes {
 
     @CoreMethod(names = "new", constructor = true, needsBlock = true, rest = true)
     public abstract static class ProcNewNode extends CoreMethodArrayArgumentsNode {
-
         public static ProcNewNode create() {
             return ProcNodesFactory.ProcNewNodeFactory.create(null);
         }
@@ -73,7 +73,8 @@ public abstract class ProcNodes {
         protected RubyProc proc(VirtualFrame frame, RubyClass procClass, Object[] args, NotProvided block,
                 @Cached FindAndReadDeclarationVariableNode readNode,
                 @Cached ReadCallerFrameNode readCaller,
-                @Cached ProcNewNode recurseNode) {
+                @Cached ProcNewNode recurseNode,
+                @Cached("new()") WarnNode warnNode) {
             final MaterializedFrame parentFrame = readCaller.execute(frame);
 
             Object parentBlock = readNode.execute(parentFrame, TranslatorEnvironment.METHOD_BLOCK_NAME, nil);
@@ -81,6 +82,12 @@ public abstract class ProcNodes {
             if (parentBlock == nil) {
                 throw new RaiseException(getContext(), coreExceptions().argumentErrorProcWithoutBlock(this));
             } else {
+                if (warnNode.shouldWarn()) {
+                    warnNode.warningMessage(
+                            getContext().getCallStack().getTopMostUserSourceSection(),
+                            "Capturing the given block using Kernel#proc is deprecated; use `&block` instead");
+                }
+
                 final RubyProc proc = (RubyProc) parentBlock;
                 return recurseNode.executeProcNew(frame, procClass, args, proc);
             }